requirements.txt
Django==4.2.11
djangorestframework==3.14.0
mysqlclient==2.2.0
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
requests==2.31.0
gunicorn==21.2.0
boto3==1.34.51
